+++ src/main/java/com/ll/exam/sbb/question/QuestionController.java
@@ -29,7 +29,7 @@ public class QuestionController
     }
 
     @RequestMapping(value = "/question/detail/{id}")
-    public String detail(Model model, @PathVariable("id") Integer id, AnswerForm answerForm) {
+    public String detail(Model model, @PathVariable("id") Long id, AnswerForm answerForm) {
         Question question = this.questionService.getQuestion(id);
         model.addAttribute("question", question);
         return "question_detail";
